Add an experimental configure script build
This build is about twice as fast as Ant and doesn't require anything
beyond the basic JDK (e.g. neither Ant nor Ivy). It can generate a .jar
file and documentation, and it can run the program in place.

I removed Checkstyle, hotswapping, and similar extra cruft that I would
never use anymore anyway. This would also be difficult to build on
Windows, which lacks the standard POSIX tools.